Adding more ability
Moved initial variables around.
Changed the first loop to include 1.
Moved the writing to CSV to the end to have the summation of the differences between the prime numbers instead of the actual prime numbers.
Made a loop to calculate the difference between prime numbers.
Made a loop to count up and transpose the quantity of each difference.